+++ CDAKit/health-data-standards/lib/import/cda/cda_allergy_importer.swift
@@ -11,8 +11,10 @@ import Fuzi
 
 class CDAKImport_CDA_AllergyImporter: CDAKImport_CDA_SectionImporter {
   var type_xpath = "./cda:code"
-  var reaction_xpath = "./cda:entryRelationship[@typeCode='MFST']/cda:observation[cda:templateId/@root='2.16.840.1.113883.10.20.1.54']/cda:value"
-  var severity_xpath = "./cda:entryRelationship[@typeCode='SUBJ']/cda:observation[cda:templateId/@root='2.16.840.1.113883.10.20.1.55']/cda:value"
+//  var reaction_xpath = "./cda:entryRelationship[@typeCode='MFST']/cda:observation[cda:templateId/@root='2.16.840.1.113883.10.20.1.54']/cda:value"
+//  var severity_xpath = "./cda:entryRelationship[@typeCode='SUBJ']/cda:observation[cda:templateId/@root='2.16.840.1.113883.10.20.1.55']/cda:value"
+  var reaction_xpath = "./cda:entryRelationship[@typeCode='MFST']/cda:observation[cda:templateId/@root='2.16.840.1.113883.10.20.1.54']"
+  var severity_xpath = "./cda:entryRelationship[@typeCode='SUBJ']/cda:observation[cda:templateId/@root='2.16.840.1.113883.10.20.1.55']"
   
   
   override init(entry_finder: CDAKImport_CDA_EntryFinder = CDAKImport_CDA_EntryFinder(entry_xpath: "//cda:observation[cda:templateId/@root='2.16.840.1.113883.10.20.1.18']")) {
@@ -35,8 +37,16 @@ class CDAKImport_CDA_AllergyImporter: CDAKImport_CDA_SectionImporter {
       extract_negation(entry_element, entry: allergy)
 
       allergy.type.addCodes(CDAKImport_CDA_SectionImporter.extract_code(entry_element, code_xpath: type_xpath))
-      allergy.reaction.addCodes(CDAKImport_CDA_SectionImporter.extract_code(entry_element, code_xpath: reaction_xpath))
-      allergy.severity.addCodes(CDAKImport_CDA_SectionImporter.extract_code(entry_element, code_xpath: severity_xpath))
+      
+//      allergy.reaction.addCodes(CDAKImport_CDA_SectionImporter.extract_code(entry_element, code_xpath: reaction_xpath))
+//      allergy.severity.addCodes(CDAKImport_CDA_SectionImporter.extract_code(entry_element, code_xpath: severity_xpath))
+      allergy.reaction = extract_entry_detail(entry_element, xpath: reaction_xpath)
+      allergy.reaction?.codes.preferred_code_sets = ["SNOMED-CT"]
+      allergy.severity = extract_entry_detail(entry_element, xpath: severity_xpath)
+      allergy.severity?.codes.preferred_code_sets = ["SNOMED-CT"]
+      
+      //extract_entry_detail
+      print(allergy.reaction)
       
       return allergy
     }