+++ CDAKit/health-data-standards/lib/models/thing_with_identifier.swift
@@ -1,0 +1,13 @@
+//
+//  thing_with_times.swift
+//  CDAKit
+//
+//  Created by Eric Whitley on 1/13/16.
+//  Copyright © 2016 Eric Whitley. All rights reserved.
+//
+
+import Foundation
+
+public protocol CDAKThingWithIdentifier: class {
+  var cda_identifier: CDAKCDAIdentifier? { get set }
+}